+++ src/primitives/link.tsx
@@ -9,6 +9,7 @@ import type { CaspianLinkProps } from './types';
 export function DefaultCaspianLink({
   href,
   className,
+  style,
   children,
   onClick,
   target,
@@ -19,6 +20,7 @@ export function DefaultCaspianLink({
     <a
       href={href}
       className={className}
+      style={style}
       onClick={onClick}
       target={target}
       rel={rel}